avances para ver tracking y detalles de pedidos ya hechos TODO: arreglar los modelos de los productos confirmados para incluir los productos, precios totales, etc nota: extra fix, productos ahora son fetch del servidor
